Abstract Data
Unhide the category/tag selector; there's a filter that allows us to force all posts to be tweeted by a specific user. Abstract Twitter OAuth keys so they're not hard-coded. If the constants are set (i.e. by a theme) use the constants, otherwise display entry fields on the existing settings page.